Use Update instead of Patch to update ownership labels on hardware resources
This addresses a hardware assignment race among reconciler threads.
The race would cause the same hardware/server to get assigned to
different tinkerbell machines.

By using Update(replace) instead of patch, we ensure that ownership
label udpate on the hardware resource happens atomically. Update
helps ensure that no changes were made to a resource between the
time it was read and the time it gets updated.

Signed-off-by: Pooja Trivedi <tripooja@amazon.com>