Do Rufio/BMC actions before creating a workflow:
This is to avoid a race condition that could occur
if a machine is running HookOS and Tink worker is
connected and running when a workflow is created.

The workflow would start to execute but then be cut
off when the reboot action is called. This has been
show to cause undefined behavior.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>