Reorganize the pkg and file structure:
Packages in Go should generally be singular
in name. So the pkg `controllers` was renamed to
`controller`. The 2 controllers, tinkerbellcluster
and tinkerbellmachine were put into their own packages.
This was done to help improve the cognitive load for
understanding and maintaining this repo.

Additional files in the machine package were created
to also help with the cognitive load of that controller.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>