Use annotations for storing provisioning state:
The previous method of provisioning state
relied on the AllowPXE field of the Hardware spec.
This overloads this field and makes it difficult
to understand state. This is also problematic when
gitops is employed with Hardware objects.

Annotations are purpose built for state and work
well with CAPI. FYI, fields in the status object state
are not maintained during CAPI move/pivot operations.
There is also at least one other known bare metal CAPI
provider that uses this pattern very successfully.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>