Add bootOptions to TinkerbellMachine spec:
Also, update to latest Workflow spec. This latest spec
will handle network and iso booting. It will also
toggle the allowPXE field in the Hardware object.

Remove using the allowPXE field to track if Hardware
is ready. Hardware is ready when the workflow is successful
and/or the Hardware has label "v1alpha1.tinkerbell.org/inUse"
equal to true.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>