Allow tinkerbellmachine objects to be removed:
When a Hardware object corresponding to a tinkerbellmachine
object no longer exists, allow the tinkerbellmachine object
to be removed and clean up associated template and workflow.
At the moment this will skip any BMC operations to power the
machine off and modify the Hardware object. This allows the
tinkerbellmachine to be deleted without needing to know that
it's ok to remove the finalizer and delete the object manually.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>