Default hardware MAC in iso mount workflows
For ISO bootmode, the controller required a placeHolder ":macAddress" to
be present in the ISOURL which would then later be replaced to the MAC
address of the actual hardware during the workflow creation time.
Default the logic to append hardware MAC before the ISO file name by
default without any placeholder, as we are doing it anyways right now
and this slightly alleviates the user experience.

Signed-off-by: Rahul Ganesh <rahulgab@amazon.com>